+++ shiori_tests/tests/vars.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+username = 'guest'
+email = 'guest@example.org'
+password = 'passw0rd'
+category = 'dummy-category'
+category2 = 'dummy2-category'
+tag = 'dummy-tag'
+tag2 = 'dummy2-tag'
+url = 'http://dummy.example.org'
+title = 'dummy title'
+description = 'This description is dummy sentense.'